+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Note: This is to facilitate larger lists with less penealty to performance when 
 * Changed anti-spam to base 7 days and up to 6 months.
 Note: Go into settings to make sure it is correct.
 * Added anti-spam for when not forced by Guild Master.
+* Added keybindings for inviting and scanning (Found in settings, Invite Settings).
+Note: This does not overwrite any existing keybindings in WoW.
 
 
 ## 2.1.36